+++ src/commands/index.ts
@@ -1,5 +1,6 @@
 import type { ExtensionModule } from '~/modules'
 import { flatten } from 'lodash'
+import checkStaleTranslations from './checkStaleTranslations'
 import configLanguages from './configLanguages'
 import configLocales from './configLocalePaths'
 import deepl from './deepl'
@@ -13,6 +14,8 @@ import keyManipulations from './keyManipulations'
 import editor from './openEditor'
 import refreshUsageReport from './refreshUsageReport'
 import review from './review'
+import scanAndExtractAll from './scanAndExtractAll'
+import translateAllMissing from './translateAllMissing'
 
 const m: ExtensionModule = (ctx) => {
   return flatten([
@@ -29,6 +32,9 @@ const m: ExtensionModule = (ctx) => {
     deepl(ctx),
     gotoRange(ctx),
     gotoNextUsage(ctx),
+    translateAllMissing(ctx),
+    scanAndExtractAll(ctx),
+    checkStaleTranslations(ctx),
   ])
 }